+++ src/App.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 
-import Router from './Router';
+import Router from './routers/Router';
 
 const App = () => {
   const queryClient = new QueryClient();
