+++ src/utils/checkAuthenticate.ts
@@ -1,0 +1,6 @@
+const checkAuthenticate = () => {
+  const accessToken = localStorage.getItem('accessToken');
+  return accessToken;
+};
+
+export default checkAuthenticate;
